// src/pages/features/TruckLoading/types.ts
export type Carregamento = {
  id: string
  scheduledAt: string
  truck: string
  driver: string
  farm: string
  field: string
  product: string
  quantity: string
  unit: string
  destination: string
}

export type CarregamentoForm = {
  scheduledAt: string
  truck: string
  driver: string
  farm: string
  field: string
  product: string
  quantity: string
  unit: string
  destination: string
}